Move route monitoring to split_tunnel module
