+++ talpid-core/src/tunnel_state_machine/connecting_state.rs
@@ -3,8 +3,6 @@ use super::{
     EventConsequence, EventResult, SharedTunnelStateValues, TunnelCommand, TunnelCommandReceiver,
     TunnelState, TunnelStateTransition, TunnelStateWrapper,
 };
-#[cfg(windows)]
-use crate::split_tunnel;
 use crate::{
     firewall::FirewallPolicy,
     routing::RouteManager,
@@ -19,8 +17,6 @@ use futures::{
     FutureExt, StreamExt,
 };
 use log::{debug, error, info, trace, warn};
-#[cfg(windows)]
-use std::ffi::OsStr;
 use std::{
     path::{Path, PathBuf},
     thread,
@@ -32,8 +28,6 @@ use talpid_types::{
     ErrorExt,
 };
 
-#[cfg(windows)]
-use super::windows;
 #[cfg(windows)]
 use crate::{routing, winnet};
 
@@ -93,18 +87,6 @@ impl ConnectingState {
                     _ => FirewallPolicyError::Generic,
                 }
             })
-    }
-
-    #[cfg(windows)]
-    fn apply_split_tunnel_config<T: AsRef<OsStr>>(
-        shared_values: &SharedTunnelStateValues,
-        paths: &[T],
-    ) -> Result<(), split_tunnel::Error> {
-        let split_tunnel = shared_values
-            .split_tunnel
-            .lock()
-            .expect("Thread unexpectedly panicked while holding the mutex");
-        split_tunnel.set_paths(paths)
     }
 
     fn start_tunnel(
@@ -332,7 +314,7 @@ impl ConnectingState {
             }
             #[cfg(windows)]
             Some(TunnelCommand::SetExcludedApps(result_tx, paths)) => {
-                let _ = result_tx.send(Self::apply_split_tunnel_config(shared_values, &paths));
+                let _ = result_tx.send(shared_values.split_tunnel.set_paths(&paths));
                 SameState(self.into())
             }
         }
@@ -459,7 +441,7 @@ impl TunnelState for ConnectingState {
             }
             Ok(tunnel_parameters) => {
                 #[cfg(windows)]
-                if let Err(error) = windows::update_split_tunnel_addresses(None, shared_values) {
+                if let Err(error) = shared_values.split_tunnel.set_tunnel_addresses(None) {
                     log::error!(
                         "{}",
                         error.display_chain_with_msg(
